requirements.txt
# requirements.txt
flask==2.3.3
torch==2.0.1
transformers==4.30.2
diffusers==0.19.3
wikipedia==1.4.0
Pillow==10.0.0
numpy==1.24.3
requests==2.31.0
pypdf2==3.0.1
accelerate==0.21.0
